project_finalFour_repo/requirements.txt
instagrapi
praw
requests
Pillow
python-dotenv
